+++ src/core/ScriptStep_prompt.ts
@@ -4,6 +4,7 @@ import { ScriptStep_Iface } from './ScriptStep_Iface'
 import type { ComfyPromptJSON } from './ComfyPrompt'
 import type { ComfyNode } from './ComfyNode'
 import { makeAutoObservable } from 'mobx'
+import { toast } from 'react-toastify'
 
 export class ScriptStep_prompt implements ScriptStep_Iface {
     name = 'prompt'
@@ -33,9 +34,19 @@ export class ScriptStep_prompt implements ScriptStep_Iface {
         }
     }
 
+    notifyEmptyPrompt = () => toast('No work to do', { type: 'warning' })
+
     /** update pointer to the currently executing node */
     onExecuting = (msg: WsMsgExecuting) => {
-        if (msg.data.node == null) return // 🔴 @comfy: why is that null sometimes ?
+        if (msg.data.node == null) {
+            console.log('🔴 null node')
+            // 2023-03-18 rvion: if I understand correctly,
+            // null here means there is no work to do.
+            if (this.currentExecutingNode == null) this.notifyEmptyPrompt()
+            this.currentExecutingNode = null
+            this._finish()
+            return
+        }
         const graph = this.execution.graph
         const node = graph.getNodeOrCrash(msg.data.node)
         this.currentExecutingNode = node
@@ -54,6 +65,11 @@ export class ScriptStep_prompt implements ScriptStep_Iface {
         // this.currentStep++
         node.artifacts.push(msg.data.output)
         console.log(node.artifacts)
+        this._finish()
+    }
+
+    /** finish this step */
+    private _finish = () => {
         if (this._resolve == null) throw new Error('❌ invariant violation: ScriptStep_prompt.resolve is null.')
         this._resolve(this)
     }