+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 2023-03-18
 
+-   🆕 NEW UI: add warning toast on empty prompt
 -   🆕 NEW UI: word wrap button in toolbar
 -   🆕 NEW core: new `ScriptExecution` context
 -   🆕 NEW placeholder documentaion website at https://rvion.github.io/StableIDE/